refactor: 공통 컴포넌트 분리(OpenGraph, 구독 유도)
