+++ src/app/articles/detail/[slug]/_components/content/ArticleContent.tsx
@@ -1,17 +1,9 @@
 'use client';
 
-import { lazy, Suspense, useEffect, useState } from 'react';
-import { NewsInfo as INewsInfo } from '@/models/article.model';
-
 import styled from 'styled-components';
-import Text from '@/components/common/Text';
-import SubscribeInduce from '@/app/articles/detail/[slug]/_components/content/SubscribeInduce';
+import SubscribeInduce from '@/components/common/article/SubscribeInduce';
 import HeightAutoImg from '@/components/common/HeightAutoImg';
-import Skeleton from '@/components/common/loader/Skeleton';
-import { getRelatedNews } from '@/hooks/useArticle';
-import { dateFormatter } from '@/utils/formatter';
-import { FaLink } from 'react-icons/fa';
-const LazyCard = lazy(() => import('@/components/common/Card'));
+import OpenGraphCard from '@/components/common/article/OpenGraphCard';
 
 interface NewsletterContentProps {
 	content: string;
@@ -23,19 +15,6 @@ interface NewsletterContentProps {
 }
 
 function ArticleContent({ content, flex, className, articleImage, related = [] }: NewsletterContentProps) {
-	const [relatedNewsData, setRelatedNewsData] = useState<INewsInfo[]>([]);
-
-	// 비동기 데이터를 useEffect에서 처리
-	useEffect(() => {
-		const fetchData = async () => {
-			if (related.length > 0) {
-				const data = await getRelatedNews(related);
-				setRelatedNewsData(data); // 상태 업데이트
-			}
-		};
-		fetchData();
-	}, [related]);
-
 	return (
 		<ContainerStyled flex={flex} className={className}>
 			{articleImage && (
@@ -48,40 +27,8 @@ function ArticleContent({ content, flex, className, articleImage, related = [] }
 				<div className="content-bottom-margin" />
 			</NewsletterContentStyled>
 			<hr />
-			<Suspense fallback={<Skeleton />}>
-				<div className="related-news">
-					<div className="header">
-						<FaLink />
-						<Text size="medium" weight="semiBold">
-							관련 뉴스
-						</Text>
-					</div>
-					{relatedNewsData.map((news) => (
-						<LazyCard
-							className="card"
-							type="list"
-							newTab={true}
-							key={news.link}
-							data={{
-								id: news.id,
-								url: news.link,
-								image: news.images[0],
-								header: news.category[0],
-								main: {
-									title: news.title,
-									description: news.content,
-								},
-								footer: (
-									<>
-										<Text color="subText">{dateFormatter(news.publishedAt)}</Text>
-									</>
-								),
-							}}
-						/>
-					))}
-				</div>
-			</Suspense>
-			<SubscribeInduce />
+			<OpenGraphCard urls={related} />
+			<SubscribeInduce className="subscribe" />
 		</ContainerStyled>
 	);
 }
@@ -98,60 +45,6 @@ const ContainerStyled = styled.div<Omit<NewsletterContentProps, 'content'>>`
 		margin: 0;
 		border: none;
 		border-bottom: 1px solid ${({ theme }) => theme.color.border};
-	}
-
-	.related-news {
-		display: flex;
-		flex-wrap: wrap;
-		gap: 1rem;
-		margin-top: 2rem;
-		margin-bottom: 4rem;
-
-		.header {
-			display: flex;
-			flex-direction: row;
-			align-items: center;
-			margin: 0;
-			padding: 0.5rem 1rem;
-			gap: 0.5rem;
-
-			border-radius: ${({ theme }) => theme.borderRadius.capsule};
-			background: ${({ theme }) => theme.color.tertiary};
-			color: ${({ theme }) => theme.color.primary};
-		}
-
-		.card {
-			background: ${({ theme }) => theme.color.surface};
-			border: 1px solid ${({ theme }) => theme.color.border};
-			border-radius: ${({ theme }) => theme.borderRadius.soft};
-			padding: 1rem;
-			gap: 0;
-
-			.card-header {
-				display: none;
-			}
-
-			.card-body {
-				border: none;
-				margin: 0;
-				padding: 0;
-				display: flex;
-				flex-direction: column;
-				justify-content: flex-start;
-
-				.content {
-					padding: 0;
-					margin: 0;
-					gap: 1rem;
-					border: none;
-
-					.left {
-						padding: 0;
-						margin: 0;
-					}
-				}
-			}
-		}
 	}
 `;
